feat: choose guide based on release when multiple guides exist for txn
When reading EDI, if multiple guides exist for a given transaction set,
resolve the correct guide by matching the envelope release with the
guide release.

When writing EDI, allow user to specify a release. Bootstrap will select
the guide for writing based on the release if multiple guides for the
transaction set are found.

Safe to upgrade.

When reading EDI, If existing users have a guide which does not match
the release of the EDI they are receiving then it will continue to
select the same guide. Logic only changes when there are multiple guides
for the same transaction set, which would previously generate an error.

When writing EDI, users must pass in a new "release" parameter to select
a guide when there are multiple guides for the same transaction set. An
error is generated if a guide is not found with the specified release.
Existing users will not pass in this parameter and will be unaffected.