feat: add configuration to send execution and file errors to destination
Execution failures occurring in bootstrap, such as destination failures,
can now be forwarded to a function, webhook, bucket, or stash. When the
error is from a destination failure it will contain information needed
to re-drive the payload to the destination.

Bind File error events from engine by default to a function. File error
events can be forwarded to a function, webhook, bucket, or stash.