feat: add execution error destination
Allows a stash configuration value to define destinations when a
processing exception is encountered. Any function failure will result
in the destinations being called.

Use the key: 'destinations|errors|execution' in the Partner
Configuration keyspace with a value of a Destination type.

Supported destinations for execution errors are: Webhook, Function,
Bucket, and Stash.

Add additional destination type of Stash.