feat: include descriptive error name and details for schema parse errors
When a user sets a configuration in Stash that is not valid against the
schema, return an error message that maximises the chance the
user can understand and resolve the issue themselves. Sets an exception
name that is more descriptive (either a stash issue or a function event
issue), and sets the context.details property to the zod issues array,
which contains the specific validation errors. Include either the stash
key name or the function name in the error message.