feat: allow destination filtering based on release
Destinations can specify a 'release' property. An inbound or outbound
transaction will only be sent to a destination if the payload (inbound
envelope or outbound metadata) release matches the destination release.
If the destination does not specify a release then it will receive
all matching transaction sets.

This follows the existing pattern for usage indicator code.