feat: add strict mode for mapping validation
The configuration options for transaction destinations is updated to
accept a 'mappingValidation' property with the value 'strict'. In order
for customers to be successful in updating configurations, the stash
item documentation was refactored to use json schemas. In addition,
the stash items themselves can now have a $schema property for better
linking back to the documentation.