fix: do not delete inbound edi file when source is core artifact bucket
when using Core's remote sftp poller, the poller adds the document to
Core's artifact bucket and sets this as the input when sending
transaction processed events. In these cases the edi-inbound function
should not attempt to delete the input since Core handles the work
and the input is read-only.

Fix tests which use the artifact bucket as the source input by default.